revert zink adition
Why?

After testing extensively, setting LIBGLE_KOPPER_DRI2=1 resolves the issue with devices like the Steam Deck. These devices have specific drivers that break when using Zink with DRI3, causing the driver to attempt to load and fail repeatedly.

By not using DRI3, most of Zink's performance advantages are undone, so fixing this issue preserves the majority of Zink's performance benefits.

I think the best approach would be to make Zink an additional setting, allowing users to choose whether to use it. This is because we can't predict if a user's device has a driver that breaks Zink with DRI3. Additionally, we could add another setting for DRI2 for users who want to use Zink on their device even if it doesn't support Zink + DRI3.

I don't think that I'm going to be able to implement it due to my lack of experience in C/C++.