Ups had realy incorrect identifier set
So just read through more documentation and noticed that i had completely misunderstood the nameing scheme so now its fixed 
if ure interested for how the string is constructed. Its basicly in oure case io.github because github is the code host url and .openstarbound.openstarbound because thats the location so basicly the identifier is just the url in a difrent format.